Finished Functionality - Save Edit & Delete are all working. Need code clean up and global error handler.

+++ BWDatabase/UI.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Runtime.Remoting.Messaging;
+using System.Linq.Expressions;
 using System.Windows.Forms;
 
 namespace BWDatabase
@@ -36,7 +35,8 @@ namespace BWDatabase
         {
             InitializeComponent();
         }
-
+ 
+        /// Userform Actions.
         private void btnClose_Click(object sender, EventArgs e)
         {
             this.Close();
@@ -45,6 +45,8 @@ namespace BWDatabase
         public void btnEdit_Click(object sender, EventArgs e)
         {
             btnEdit.Enabled = false;
+            btnCreate.Enabled = false;
+            btnDelete.Enabled = false;
             btnEditSave.Visible = true;
             btnAllCancel.Visible = true;
             lstCustomers.Enabled = false;
@@ -63,7 +65,6 @@ namespace BWDatabase
             gNotes = txtNotes.Text;
             //
             LockOrUnlockFields(true);
-
         }
 
         private void frmDatabase_Load(object sender, EventArgs e)
@@ -74,84 +75,23 @@ namespace BWDatabase
 
         private void lstCustomers_SelectedIndexChanged(object sender, EventArgs e)
         {
-            if(lstCustomers.SelectedItem.ToString()== NewCust)
+            try
             {
-
-            }
-            else
-            {
-            ClearCustomerValues();
-            LoadCustomerValues();
-            }
-
-        }
-        public void LoadCustomerList()
-        {
-            Database db = new BWDatabase.Database();
-
-            {
-                foreach (var Row in db.GetCustomerList())
+                if (lstCustomers.SelectedItem.ToString() == NewCust)
                 {
-                    var Fields = Row as IDictionary<string, object>;
-                    FName = Fields["FullName"].ToString();
-                    this.lstCustomers.Items.Add(FName);
+                    //Do Nothing
                 }
-            }
-        }
-
-        public void LoadCustomerValues()
-        {
-            Database db = new BWDatabase.Database();
-            {
-                string getValue = lstCustomers.SelectedItem.ToString();
-                foreach (var Row in db.GetCustomerDetails(getValue))
+                else
                 {
-                    var Fields = Row as IDictionary<string, object>;
-                    txtCustNumber.Text = Fields["CustomerNumber"].ToString();
-                    txtTitle.Text = Fields["Title"].ToString();
-                    txtFullName.Text = Fields["FullName"].ToString();
-                    txtTelephone.Text = Fields["Telephone"].ToString();
-                    txtAddTelephone.Text = Fields["AddTelephone"].ToString();
-                    txtMobile.Text = Fields["Mobile"].ToString();
-                    txtEmail.Text = Fields["Email"].ToString();
-                    txtFullAddress.Text = Fields["FullAddress"].ToString();
-                    txtPostcode.Text = Fields["Postcode"].ToString();
-                    txtArea.Text = Fields["Area"].ToString();
-                    txtFacebookName.Text = Fields["Facebook"].ToString();
-                    txtLanguage.Text = Fields["Language"].ToString();
-                    txtNotes.Text = Fields["Notes"].ToString();
-                }
-            }
-            if(txtTelephone.Text == "0")
+                    ClearCustomerValues();
+                    LoadCustomerValues();
+                }               
+            }     
+            catch(NullReferenceException)
             {
-                txtTelephone.Text = "";
+                //Cannot replicate previous issue where null is handed as the index.
+                lstCustomers.SelectedItem = 0;
             }
-            if (txtAddTelephone.Text == "0")
-            {
-                txtAddTelephone.Text = "";
-            }
-            if (txtMobile.Text == "0")
-            {
-                txtMobile.Text = "";
-            }
-
-        }
-
-        public void ClearCustomerValues()
-        {
-            txtCustNumber.Text = null;
-            txtTitle.Text = null;
-            txtFullName.Text = null;
-            txtTelephone.Text = null;
-            txtAddTelephone.Text = null;
-            txtMobile.Text = null;
-            txtEmail.Text = null;
-            txtFullAddress.Text = null;
-            txtPostcode.Text = null;
-            txtArea.Text = null;
-            txtFacebookName.Text = null;
-            txtLanguage.Text = null;
-            txtNotes.Text = null;
         }
 
         private void btnCreate_Click(object sender, EventArgs e)
@@ -200,13 +140,18 @@ namespace BWDatabase
                                 SaveNewEntry(CustomerNumber);
                                 ClearCustomerList();
                                 LoadCustomerList();
-                                lstCustomers.SelectedIndex = 0;
                                 lstCustomers.Enabled = true;
                                 btnCreate.Text = CreateNew;
                                 btnEdit.Enabled = true;
                                 btnDelete.Enabled = true;
                                 btnAllCancel.Visible = false;
                                 LockOrUnlockFields(false);
+
+                                string Identifier = gFullName;
+                                lstCustomers.SelectedIndex = FindIndex(Identifier);
+
+                                System.Threading.Thread.Sleep(100);
+
                                 break;
                             }
                             else
@@ -230,47 +175,22 @@ namespace BWDatabase
                     }                    
             }
         }
-
-        public void AddDefaultClearValues()
-        {
-            lstCustomers.Items.Add(NewCust);
-            int Count = lstCustomers.Items.Count;
-            lstCustomers.SelectedIndex = Count - 1;
-        }
-
-        public void SaveNewEntry(int Count)
-        {
-            gTitle = txtTitle.Text;
-            gFullName = txtFullName.Text;
-            gTelephone = txtTelephone.Text;
-            gAddTelephone = txtAddTelephone.Text;
-            gMobile = txtMobile.Text;
-            gEmail = txtEmail.Text;
-            gFullAddress = txtFullAddress.Text;
-            gPostCode = txtPostcode.Text;
-            gArea = txtArea.Text;
-            gFacebook = txtFacebookName.Text;
-            gLanguage = txtLanguage.Text;
-            gNotes = txtNotes.Text;
-            
-            Database db = new BWDatabase.Database();
-            db.SaveEntry(Count, gTitle, gFullName, gTelephone, gAddTelephone, gMobile, gEmail, 
-                gFullAddress, gPostCode, gArea, gFacebook, gLanguage, gNotes);
-        }
-
-        public void ClearCustomerList()
-        {
-            lstCustomers.Items.Clear();
-        }
-
+    
         private void btnDelete_Click(object sender, EventArgs e)
-        {
-            Database db = new BWDatabase.Database();
-            CustNumber = Int32.Parse(txtCustNumber.Text);
-            db.DeleteEntry(CustNumber, txtFullName.Text);
-            ClearCustomerList();
-            ClearCustomerValues();
-            LoadCustomerList();
+        {                
+            if(lstCustomers.SelectedIndex < 0)
+            {
+                MessageBox.Show("No Selection.");
+            }
+            else
+            {
+                Database db = new BWDatabase.Database();
+                CustNumber = Int32.Parse(txtCustNumber.Text);
+                db.DeleteEntry(CustNumber, txtFullName.Text);
+                ClearCustomerList();
+                ClearCustomerValues();
+                LoadCustomerList();
+            }
         }
 
         private void btnAllCancel_Click(object sender, EventArgs e)
@@ -284,6 +204,7 @@ namespace BWDatabase
                 lstCustomers.SelectedIndex = 0;
                 btnAllCancel.Visible = false;
                 btnEdit.Enabled = true;
+                btnDelete.Enabled = true;
                 LockOrUnlockFields(false);
             }
             else
@@ -291,6 +212,8 @@ namespace BWDatabase
                 btnEdit.Enabled = true;
                 btnEditSave.Visible = false;
                 btnAllCancel.Visible = false;
+                btnCreate.Enabled = true;
+                btnDelete.Enabled = true;
                 lstCustomers.Enabled = true;
                 txtTitle.Text = gTitle;
                 txtFullName.Text = gFullName;
@@ -311,47 +234,76 @@ namespace BWDatabase
 
         private void btnEditSave_Click(object sender, EventArgs e)
         {
+            int num;
+            num = Int32.Parse(txtCustNumber.Text);
+            EditExistingEntry(num);
+            LockOrUnlockFields(false);
+            btnEditSave.Visible = false;
+            btnAllCancel.Visible = false;
+            lstCustomers.Enabled = true;
+            btnCreate.Enabled = true;
+            btnEdit.Enabled = true;
+            btnDelete.Enabled = true;
+            ClearCustomerList();
+            LoadCustomerList();
 
+            string Identifier = gFullName;
+            lstCustomers.SelectedIndex = FindIndex(Identifier);
         }
-
+/// Manipulation of the Userform.
         private void LockOrUnlockFields(Boolean LockorUnlock)
         {
             switch(LockorUnlock)
             {
-                case true:
-                    {
-                        txtTitle.ReadOnly = false;
-                        txtFullName.ReadOnly = false;
-                        txtTelephone.ReadOnly = false;
-                        txtAddTelephone.ReadOnly = false;
-                        txtMobile.ReadOnly = false;
-                        txtEmail.ReadOnly = false;
-                        txtFullAddress.ReadOnly = false;
-                        txtPostcode.ReadOnly = false;
-                        txtArea.ReadOnly = false;
-                        txtFacebookName.ReadOnly = false;
-                        txtLanguage.ReadOnly = false;
-                        txtNotes.ReadOnly = false;
-                        break;
-                    }
-                case false:
-                    {
-                        txtTitle.ReadOnly = true;
-                        txtFullName.ReadOnly = true;
-                        txtTelephone.ReadOnly = true;
-                        txtAddTelephone.ReadOnly = true;
-                        txtMobile.ReadOnly = true;
-                        txtEmail.ReadOnly = true;
-                        txtFullAddress.ReadOnly = true;
-                        txtPostcode.ReadOnly = true;
-                        txtArea.ReadOnly = true;
-                        txtFacebookName.ReadOnly = true;
-                        txtLanguage.ReadOnly = true;
-                        txtNotes.ReadOnly = true;
-                        break;
-                    }
+                case true:                
+                    txtTitle.ReadOnly = false;
+                    txtFullName.ReadOnly = false;
+                    txtTelephone.ReadOnly = false;
+                    txtAddTelephone.ReadOnly = false;
+                    txtMobile.ReadOnly = false;
+                    txtEmail.ReadOnly = false;
+                    txtFullAddress.ReadOnly = false;
+                    txtPostcode.ReadOnly = false;
+                    txtArea.ReadOnly = false;
+                    txtFacebookName.ReadOnly = false;
+                    txtLanguage.ReadOnly = false;
+                    txtNotes.ReadOnly = false;
+                    break;
+                    
+                case false:                    
+                    txtTitle.ReadOnly = true;
+                    txtFullName.ReadOnly = true;
+                    txtTelephone.ReadOnly = true;
+                    txtAddTelephone.ReadOnly = true;
+                    txtMobile.ReadOnly = true;
+                    txtEmail.ReadOnly = true;
+                    txtFullAddress.ReadOnly = true;
+                    txtPostcode.ReadOnly = true;
+                    txtArea.ReadOnly = true;
+                    txtFacebookName.ReadOnly = true;
+                    txtLanguage.ReadOnly = true;
+                    txtNotes.ReadOnly = true;
+                    break;                    
             }
         }
+
+        public int FindIndex(string Identifier)
+        {           
+        int i;
+        int lstCount = lstCustomers.Items.Count;
+        for (i = 0; i < lstCount; i ++)
+            {
+                string CustName = lstCustomers.Items[i].ToString();
+                if(CustName == Identifier)
+                {
+                    Console.WriteLine(i);
+                    break;
+                }                
+            }
+            Console.WriteLine(i);
+            return i;
+        }
+/// Validation methods.
         public bool ValidateContactFields()
         {
             foreach (Control control in grpContact.Controls)
@@ -406,6 +358,124 @@ namespace BWDatabase
                 }
             }
             return true;
+        }
+/// Manipulation of the information held within the list.
+        public void LoadCustomerList()
+        {
+            Database db = new BWDatabase.Database();
+
+            {
+                foreach (var Row in db.GetCustomerList())
+                {
+                    var Fields = Row as IDictionary<string, object>;
+                    FName = Fields["FullName"].ToString();
+                    this.lstCustomers.Items.Add(FName);
+                }
+            }
+        }
+
+        public void ClearCustomerList()
+        {
+            lstCustomers.Items.Clear();
+        }
+       
+        public void LoadCustomerValues()
+        {
+            Database db = new BWDatabase.Database();
+            {
+                string getValue = lstCustomers.SelectedItem.ToString();
+                foreach (var Row in db.GetCustomerDetails(getValue))
+                {
+                    var Fields = Row as IDictionary<string, object>;
+                    txtCustNumber.Text = Fields["CustomerNumber"].ToString();
+                    txtTitle.Text = Fields["Title"].ToString();
+                    txtFullName.Text = Fields["FullName"].ToString();
+                    txtTelephone.Text = Fields["Telephone"].ToString();
+                    txtAddTelephone.Text = Fields["AddTelephone"].ToString();
+                    txtMobile.Text = Fields["Mobile"].ToString();
+                    txtEmail.Text = Fields["Email"].ToString();
+                    txtFullAddress.Text = Fields["FullAddress"].ToString();
+                    txtPostcode.Text = Fields["Postcode"].ToString();
+                    txtArea.Text = Fields["Area"].ToString();
+                    txtFacebookName.Text = Fields["Facebook"].ToString();
+                    txtLanguage.Text = Fields["Language"].ToString();
+                    txtNotes.Text = Fields["Notes"].ToString();
+                }
+            }
+            if (txtTelephone.Text == "0")
+            {
+                txtTelephone.Text = "";
+            }
+            if (txtAddTelephone.Text == "0")
+            {
+                txtAddTelephone.Text = "";
+            }
+            if (txtMobile.Text == "0")
+            {
+                txtMobile.Text = "";
+            }
+        }
+        public void AddDefaultClearValues()
+        {
+            lstCustomers.Items.Add(NewCust);
+            int Count = lstCustomers.Items.Count;
+            lstCustomers.SelectedIndex = Count - 1;
+        }
+/// Manipulation of Data.
+        public void ClearCustomerValues()
+        {
+            txtCustNumber.Text = null;
+            txtTitle.Text = null;
+            txtFullName.Text = null;
+            txtTelephone.Text = null;
+            txtAddTelephone.Text = null;
+            txtMobile.Text = null;
+            txtEmail.Text = null;
+            txtFullAddress.Text = null;
+            txtPostcode.Text = null;
+            txtArea.Text = null;
+            txtFacebookName.Text = null;
+            txtLanguage.Text = null;
+            txtNotes.Text = null;
+        }
+
+        public void EditExistingEntry(int CustNum)
+        {
+            gTitle = txtTitle.Text;
+            gFullName = txtFullName.Text;
+            gTelephone = txtTelephone.Text;
+            gAddTelephone = txtAddTelephone.Text;
+            gMobile = txtMobile.Text;
+            gEmail = txtEmail.Text;
+            gFullAddress = txtFullAddress.Text;
+            gPostCode = txtPostcode.Text;
+            gArea = txtArea.Text;
+            gFacebook = txtFacebookName.Text;
+            gLanguage = txtLanguage.Text;
+            gNotes = txtNotes.Text;
+
+            Database db = new BWDatabase.Database();
+            db.EditEntry(CustNum, gTitle, gFullName, gTelephone, gAddTelephone, gMobile, gEmail,
+                gFullAddress, gPostCode, gArea, gFacebook, gLanguage, gNotes);
+        }
+        public void SaveNewEntry(int CustNum)
+        {
+            gTitle = txtTitle.Text;
+            gFullName = txtFullName.Text;
+            gTelephone = txtTelephone.Text;
+            gAddTelephone = txtAddTelephone.Text;
+            gMobile = txtMobile.Text;
+            gEmail = txtEmail.Text;
+            gFullAddress = txtFullAddress.Text;
+            gPostCode = txtPostcode.Text;
+            gArea = txtArea.Text;
+            gFacebook = txtFacebookName.Text;
+            gLanguage = txtLanguage.Text;
+            gNotes = txtNotes.Text;
+
+            Database db = new BWDatabase.Database();
+            db.SaveEntry(CustNum, gTitle, gFullName, gTelephone, gAddTelephone, gMobile, gEmail,
+                gFullAddress, gPostCode, gArea, gFacebook, gLanguage, gNotes);
         }
     }
 }
